+++ Asteroids/Misc Functions.cpp
@@ -23,5 +23,5 @@ std::string GetEXEPath()
 
 float GetDistance(float x1, float x2, float y1, float y2)
 {
-	return sqrt( pow(x2 - x1, 2) + pow(y2 - y1, 2) );
+	return static_cast <float> (sqrt( pow(x2 - x1, 2) + pow(y2 - y1, 2) ));
 }
